agents/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
google-generativeai==0.3.2
chromadb==0.4.18
pandas==2.1.3
numpy==1.25.2
tabulate==0.9.0
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.0